move vertical-work code into a sub-directory, fix more failing tests
