+++ packages/typesetter/src/typesetters/table.ts
@@ -73,7 +73,7 @@ export const typesetTable = (
               ];
 
     const topRowChildren = children.slice(0, node.colCount);
-    const MIN_WIDTH = 32; // only use this for debugging purposes.
+    const MIN_WIDTH = 0; // 32; // only use this for debugging purposes.
 
     // Group cells into rows and columns and determine the width of each
     // column and the depth/height of each row.
